+++ flash/ek/flash/doc/types.hpp
@@ -134,24 +134,14 @@ enum class fill_type {
     unknown = 0,
     solid = 1,
     linear = 2,
-    radial = 3
+    radial = 3,
+    bitmap = 4
 };
 
 enum class spread_method {
     extend = 0,
     reflect = 1,
     repeat = 2
-};
-
-struct solid_stroke {
-    float4 fill;
-    scale_mode scaleMode = scale_mode::none;
-    solid_style_type solidStyle = solid_style_type::hairline;
-    float weight = 1.0f;
-    line_caps caps = line_caps::round;
-    line_joints joints = line_joints::round;
-    float miterLimit = 0.0f;
-    bool pixelHinting = false;
 };
 
 struct gradient_entry {
@@ -168,16 +158,27 @@ struct gradient_entry {
 
 struct fill_style {
     int index = 0;
-    fill_type type;
-    spread_method spreadMethod;
+
+    fill_type type = fill_type::solid;
+    spread_method spreadMethod = spread_method::repeat;
     std::vector<gradient_entry> entries;
-    matrix_2d matrix;
+    matrix_2d matrix{};
+    std::string bitmapPath;
+    mutable const bitmap_t* bitmap = nullptr;
 };
 
 struct stroke_style {
     int index = 0;
-    solid_stroke solid;
-    bool is_solid = false;
+    bool is_solid = true;
+
+    fill_style fill{};
+    scale_mode scaleMode = scale_mode::none;
+    solid_style_type solidStyle = solid_style_type::hairline;
+    float weight = 1.0f;
+    line_caps caps = line_caps::round;
+    line_joints joints = line_joints::round;
+    float miterLimit = 0.0f;
+    bool pixelHinting = false;
 };
 
 /**** elements tree ****/